refactor ui app graph state
